fix nip07 signer casting
